# Add fast_float header library for fast string->double, and optimize Parquet WKT->WKB conversion