--- a/ogr/ogr_wkb.cpp
+++ b/ogr/ogr_wkb.cpp
@@ -34,6 +34,14 @@
 #include <cmath>
 #include <climits>
 
+#include <algorithm>
+#include <limits>
+
+#define USE_FAST_FLOAT
+#ifdef USE_FAST_FLOAT
+#include "include_fast_float.h"
+#endif
+
 /************************************************************************/
 /*                          OGRWKBNeedSwap()                            */
 /************************************************************************/
@@ -269,4 +277,252 @@ const GByte *WKBFromEWKB(GByte *pabyEWKB, size_t nEWKBSize, size_t &nWKBSizeOut,
     }
 
     return pabyWKB;
+}
+
+/************************************************************************/
+/*                         OGRAppendBuffer()                            */
+/************************************************************************/
+
+OGRAppendBuffer::OGRAppendBuffer() = default;
+
+/************************************************************************/
+/*                        ~OGRAppendBuffer()                            */
+/************************************************************************/
+
+OGRAppendBuffer::~OGRAppendBuffer() = default;
+
+/************************************************************************/
+/*                       OGRWKTToWKBTranslator()                        */
+/************************************************************************/
+
+OGRWKTToWKBTranslator::OGRWKTToWKBTranslator(OGRAppendBuffer &oAppendBuffer)
+    : m_oAppendBuffer(oAppendBuffer)
+{
+#ifndef USE_FAST_FLOAT
+    // Test if current locale decimal separator is decimal point
+    char szTest[10];
+    snprintf(szTest, sizeof(szTest), "%f", 1.5);
+    m_bCanUseStrtod = strchr(szTest, '.') != nullptr;
+#endif
+    CPL_IGNORE_RET_VAL(m_bCanUseStrtod);
+}
+
+/************************************************************************/
+/*                          TranslateWKT()                              */
+/************************************************************************/
+
+size_t OGRWKTToWKBTranslator::TranslateWKT(void *pabyWKTStart, size_t nLength,
+                                           bool bCanAlterByteAfter)
+{
+    const char *pszPtrStart = static_cast<const char *>(pabyWKTStart);
+    // Optimize single-part single-ring multipolygon WKT->WKB translation
+    if (bCanAlterByteAfter && nLength > strlen("MULTIPOLYGON") &&
+        EQUALN(pszPtrStart, "MULTIPOLYGON", strlen("MULTIPOLYGON")))
+    {
+        int nCountOpenPar = 0;
+        size_t nCountComma = 0;
+        bool bHasZ = false;
+        bool bHasM = false;
+
+        char *pszEnd = static_cast<char *>(pabyWKTStart) + nLength;
+        const char chBackup = *pszEnd;
+        *pszEnd = 0;
+
+        // Checks that the multipolygon consists of a single part with
+        // only an exterior ring.
+        for (const char *pszPtr = pszPtrStart + strlen("MULTIPOLYGON"); *pszPtr;
+             ++pszPtr)
+        {
+            const char ch = *pszPtr;
+            if (ch == 'Z')
+                bHasZ = true;
+            else if (ch == 'M')
+                bHasM = true;
+            if (ch == '(')
+            {
+                nCountOpenPar++;
+                if (nCountOpenPar == 4)
+                    break;
+            }
+            else if (ch == ')')
+            {
+                nCountOpenPar--;
+                if (nCountOpenPar < 0)
+                    break;
+            }
+            else if (ch == ',')
+            {
+                if (nCountOpenPar < 3)
+                {
+                    // multipart / multi-ring
+                    break;
+                }
+                nCountComma++;
+            }
+        }
+        const int nDim = 2 + (bHasZ ? 1 : 0) + (bHasM ? 1 : 0);
+        if (nCountOpenPar == 0 && nCountComma > 0 &&
+            nCountComma < std::numeric_limits<uint32_t>::max())
+        {
+            const uint32_t nVerticesCount =
+                static_cast<uint32_t>(nCountComma + 1);
+            const size_t nWKBSize =
+                sizeof(GByte) +     // Endianness
+                sizeof(uint32_t) +  // multipolygon WKB geometry type
+                sizeof(uint32_t) +  // number of parts
+                sizeof(GByte) +     // Endianness
+                sizeof(uint32_t) +  // polygon WKB geometry type
+                sizeof(uint32_t) +  // number of rings
+                sizeof(uint32_t) +  // number of vertices
+                nDim * sizeof(double) * nVerticesCount;
+            GByte *const pabyCurStart = static_cast<GByte *>(
+                m_oAppendBuffer.GetPtrForNewBytes(nWKBSize));
+            if (!pabyCurStart)
+            {
+                return static_cast<size_t>(-1);
+            }
+            GByte *pabyCur = pabyCurStart;
+            // Multipolygon byte order
+            {
+                *pabyCur = wkbNDR;
+                pabyCur++;
+            }
+            // Multipolygon geometry type
+            {
+                uint32_t nWKBGeomType =
+                    wkbMultiPolygon + (bHasZ ? 1000 : 0) + (bHasM ? 2000 : 0);
+                CPL_LSBPTR32(&nWKBGeomType);
+                memcpy(pabyCur, &nWKBGeomType, sizeof(uint32_t));
+                pabyCur += sizeof(uint32_t);
+            }
+            // Number of parts
+            {
+                uint32_t nOne = 1;
+                CPL_LSBPTR32(&nOne);
+                memcpy(pabyCur, &nOne, sizeof(uint32_t));
+                pabyCur += sizeof(uint32_t);
+            }
+            // Polygon byte order
+            {
+                *pabyCur = wkbNDR;
+                pabyCur++;
+            }
+            // Polygon geometry type
+            {
+                uint32_t nWKBGeomType =
+                    wkbPolygon + (bHasZ ? 1000 : 0) + (bHasM ? 2000 : 0);
+                CPL_LSBPTR32(&nWKBGeomType);
+                memcpy(pabyCur, &nWKBGeomType, sizeof(uint32_t));
+                pabyCur += sizeof(uint32_t);
+            }
+            // Number of rings
+            {
+                uint32_t nOne = 1;
+                CPL_LSBPTR32(&nOne);
+                memcpy(pabyCur, &nOne, sizeof(uint32_t));
+                pabyCur += sizeof(uint32_t);
+            }
+            // Number of vertices
+            {
+                uint32_t nVerticesCountToWrite = nVerticesCount;
+                CPL_LSBPTR32(&nVerticesCountToWrite);
+                memcpy(pabyCur, &nVerticesCountToWrite, sizeof(uint32_t));
+                pabyCur += sizeof(uint32_t);
+            }
+            uint32_t nDoubleCount = 0;
+            const uint32_t nExpectedDoubleCount = nVerticesCount * nDim;
+            for (const char *pszPtr = pszPtrStart + strlen("MULTIPOLYGON");
+                 *pszPtr;
+                 /* nothing */)
+            {
+                const char ch = *pszPtr;
+                if (ch == '-' || ch == '.' || (ch >= '0' && ch <= '9'))
+                {
+                    nDoubleCount++;
+                    if (nDoubleCount > nExpectedDoubleCount)
+                    {
+                        break;
+                    }
+#ifdef USE_FAST_FLOAT
+                    double dfVal;
+                    auto answer = fast_float::from_chars(pszPtr, pszEnd, dfVal);
+                    if (answer.ec != std::errc())
+                    {
+                        nDoubleCount = 0;
+                        break;
+                    }
+                    pszPtr = answer.ptr;
+#else
+                    char *endptr = nullptr;
+                    const double dfVal =
+                        m_bCanUseStrtod ? strtod(pszPtr, &endptr)
+                                        : CPLStrtodDelim(pszPtr, &endptr, '.');
+                    pszPtr = endptr;
+#endif
+                    CPL_LSBPTR64(&dfVal);
+                    memcpy(pabyCur, &dfVal, sizeof(double));
+                    pabyCur += sizeof(double);
+                }
+                else
+                {
+                    ++pszPtr;
+                }
+            }
+            if (nDoubleCount == nExpectedDoubleCount)
+            {
+                CPLAssert(static_cast<size_t>(pabyCur - pabyCurStart) ==
+                          nWKBSize);
+                // cppcheck-suppress selfAssignment
+                *pszEnd = chBackup;
+                return nWKBSize;
+            }
+            else
+            {
+                CPLError(CE_Failure, CPLE_AppDefined,
+                         "Invalid WKT geometry: %s", pszPtrStart);
+                // cppcheck-suppress selfAssignment
+                *pszEnd = chBackup;
+                return static_cast<size_t>(-1);
+            }
+        }
+        // cppcheck-suppress selfAssignment
+        *pszEnd = chBackup;
+    }
+
+    // General case going through a OGRGeometry
+    OGRGeometry *poGeometry = nullptr;
+    if (bCanAlterByteAfter)
+    {
+        // Slight optimization for all geometries but the final one, to
+        // avoid creating a new string each time.
+        // We set the ending byte to '\0' and restore it back after parsing
+        // the WKT
+        char *pszEnd = static_cast<char *>(pabyWKTStart) + nLength;
+        const char chBackup = *pszEnd;
+        *pszEnd = 0;
+        OGRGeometryFactory::createFromWkt(pszPtrStart, nullptr, &poGeometry);
+        // cppcheck-suppress selfAssignment
+        *pszEnd = chBackup;
+    }
+    else
+    {
+        std::string osTmp;
+        osTmp.assign(pszPtrStart, nLength);
+        OGRGeometryFactory::createFromWkt(osTmp.c_str(), nullptr, &poGeometry);
+    }
+    if (!poGeometry)
+    {
+        CPLError(CE_Failure, CPLE_AppDefined, "Invalid WKT geometry");
+        return static_cast<size_t>(-1);
+    }
+    const size_t nWKBSize = poGeometry->WkbSize();
+    GByte *pabyWKB =
+        static_cast<GByte *>(m_oAppendBuffer.GetPtrForNewBytes(nWKBSize));
+    if (!pabyWKB)
+    {
+        return static_cast<size_t>(-1);
+    }
+    poGeometry->exportToWkb(wkbNDR, pabyWKB, wkbVariantIso);
+    delete poGeometry;
+    return nWKBSize;
 }