… ogr/ogrsf_frmts/arrow_common/ograrrowlayer.hpp
… port/cpl_strtod.cpp
… port/cpl_virtualmem.cpp
… port/include_fast_float.h
… scripts/cppcheck.sh
… third_party/fast_float/AUTHORS
… third_party/fast_float/LICENSE-MIT
… third_party/fast_float/PROVENANCE.TXT
… third_party/fast_float/ascii_number.h
… third_party/fast_float/bigint.h
… third_party/fast_float/constexpr_feature_detect.h
… third_party/fast_float/decimal_to_binary.h
… third_party/fast_float/digit_comparison.h
… third_party/fast_float/fast_float.h
… third_party/fast_float/fast_table.h
… third_party/fast_float/float_common.h
… third_party/fast_float/parse_number.h